Pie chart refactor, tooltips

+++ src/gui/gui_common_elements.hpp
@@ -1001,8 +1001,8 @@ public:
 
 class upper_house_piechart : public piechart<dcon::ideology_id> {
 protected:
-	std::unordered_map<int32_t, float> get_distribution(sys::state& state) noexcept override {
-		std::unordered_map<int32_t, float> distrib = {};
+	std::unordered_map<uint8_t, float> get_distribution(sys::state& state) noexcept override {
+		std::unordered_map<uint8_t, float> distrib = {};
 		Cyto::Any nat_id_payload = dcon::nation_id{};
 		if(parent) {
 			parent->impl_get(state, nat_id_payload);
@@ -1010,7 +1010,7 @@ protected:
 				auto nat_id = any_cast<dcon::nation_id>(nat_id_payload);
 				state.world.for_each_ideology([&](dcon::ideology_id ideo_id) {
 					auto weight = .01f * state.world.nation_get_upper_house(nat_id, ideo_id);
-					distrib[ideo_id.index()] = weight;
+					distrib[uint8_t(ideo_id.index())] = weight;
 				});
 			}
 		}
@@ -1020,8 +1020,8 @@ protected:
 
 class voter_ideology_piechart : public piechart<dcon::ideology_id> {
 protected:
-	std::unordered_map<int32_t, float> get_distribution(sys::state& state) noexcept override {
-		std::unordered_map<int32_t, float> distrib = {};
+	std::unordered_map<uint8_t, float> get_distribution(sys::state& state) noexcept override {
+		std::unordered_map<uint8_t, float> distrib = {};
 		Cyto::Any nat_id_payload = dcon::nation_id{};
 		if(parent) {
 			parent->impl_get(state, nat_id_payload);
@@ -1048,7 +1048,7 @@ protected:
 					}
 				});
 				for(size_t i = 0; i < ideo_pool.size(); i++) {
-					distrib[i] = ideo_pool[i] / total;
+					distrib[uint8_t(i)] = ideo_pool[i] / total;
 				}
 			}
 		}
